move router functionality to new layout
