+++ projects/storefrontlib/src/cms-structure/index.ts
@@ -1,4 +1,5 @@
 export * from './outlet/index';
 export * from './page/index';
 export * from './pwa/index';
+export * from './routing/index';
 export * from './seo/index';
