+++ projects/storefrontlib/src/cms-structure/routing/cms-route/add-cms-route.ts
@@ -1,7 +1,7 @@
 import { Injector } from '@angular/core';
 import { Route, Router } from '@angular/router';
-import { PageLayoutComponent } from '../../cms-structure/page/index';
-import { CmsPageGuard } from '../cms/guards/cms-page.guard';
+import { CmsPageGuard } from '../../../lib/cms/guards/cms-page.guard';
+import { PageLayoutComponent } from '../../page/index';
 
 const cmsRoute: Route = {
   path: '**',
