+++ projects/storefrontlib/src/cms-structure/routing/index.ts
@@ -1,0 +1,2 @@
+export * from './cms-route/index';
+export * from './suffix-routes/index';
