+++ projects/storefrontlib/src/lib/storefront.module.ts
@@ -11,12 +11,12 @@ import {
   StateModule,
 } from '@spartacus/core';
 import { CmsLibModule } from '../cms-components/index';
+import { CmsRouteModule } from '../cms-structure/routing/cms-route/cms-route.module';
+import { SuffixRoutesModule } from '../cms-structure/routing/suffix-routes/suffix-routes.module';
 import { MultiStepCheckoutModule } from './checkout/index';
-import { CmsRouteModule } from './cms-route/cms-route.module';
 import { CmsModule } from './cms/index';
 import { provideConfigFromMetaTags } from './provide-config-from-meta-tags';
 import { StorefrontModuleConfig } from './storefront-config';
-import { SuffixRoutesModule } from './suffix-routes/suffix-routes.module';
 import { UiModule } from './ui/index';
 
 @NgModule({
