+++ projects/storefrontlib/src/public_api.ts
@@ -7,11 +7,9 @@ export * from './cms-components/myaccount/index';
 export * from './cms-structure/index';
 export * from './layout/index';
 export * from './lib/checkout/index';
-export * from './lib/cms-route/index';
 export * from './lib/cms/index';
 export * from './lib/storefront-config';
 export * from './lib/storefront.module';
-export * from './lib/suffix-routes/index';
 export * from './lib/ui/index';
 export * from './shared/index';
 export * from './translations/index';
